+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+
+- the structur of the userinfo
 
 ## [1.1] - 2019-10-10
 ### Added